Add empty test case to phone-number_test.hs
When solving the Haskell phone number assignment with a pattern match, one could get into a situation where the patterns are non-exhaustive for empty strings. The resulting error should be triggered by the automatic test. This commit adds another test case to assert that an empty string is treated like an invalid phone number.