pov: Add alternate expected orderings for rerooted trees
Given a tree that must be reparented on node `x`, the tests used to
ensure that the former parent of `x` comes *after* the former children
of `x`. But there's no reason this should be the case, since in trees we
don't particularly care about the ordering of the children. (There are
some specialised trees for which it is the case, but this is not one of
those trees)

This test allows either ordering (former parent before former children,
or former children before former parent).

Tested by ensuring that the example solution passes (it uses the latter
ordering) and the example solution still passes when using the former
ordering (modify `(children g) ++ [reparented]` to instead be
`reparented : children g`.

Also tested to ensure that solutions that always return `Just []` and
`Nothing` return reasonable error messages.

```
rerooted tree should be one of the possibilities:
1: Graph {value = "x", children = [Graph {value = "kid-0", children = []},Graph {value = "kid-1", children = []},Graph {value = "root", children = []}]}
2: Graph {value = "x", children = [Graph {value = "root", children = []},Graph {value = "kid-0", children = []},Graph {value = "kid-1", children = []}]}
but got: Just (Graph {value = "x", children = []})
```

Fixes #92