pov: Allow multiple possibilities
Change the test expectations to allow multiple trees in the case where a
rerooted tree should exist. If there are multiple trees, the observed
output just needs to equal one of the expected trees. This is useful
because sometimes there are multiple reasonable answers (regarding the
ordering of the former parent of the root versus the former children of
the root).