binary-search-tree: rm example redundant constraints
Note that this only affects the example, not the tests.

GHC 8.0.1 will give the error:

```
BST.hs:15:1: warning: [-Wredundant-constraints]
    • Redundant constraint: Ord a
    • In the type signature for:
           singleton :: Ord a => a -> BST a
```

That makes sense, since you can make any value a singleton tree; it
doesn't necessarily have to be Ord. Of course, you'll find yourself
unable to add other values to the tree if it's not Ord, but that's not
`singleton`'s problem.