custom-set: rm example redundant constraints
Note that this only affects the example, not the tests.

GHC 8.0.1 will give the error:

```
CustomSet.hs:30:10: warning: [-Wredundant-constraints]
    • Redundant constraint: Eq a
    • In the instance declaration for ‘Eq (CustomSet a)’

CustomSet.hs:50:1: warning: [-Wredundant-constraints]
    • Redundant constraint: Ord a
    • In the type signature for:
           singleton :: Ord a => a -> CustomSet a

CustomSet.hs:56:1: warning: [-Wredundant-constraints]
    • Redundant constraint: Ord a
    • In the type signature for:
           empty :: Ord a => CustomSet a
```

The singleton redundancy makes sense, since you can make any value a
singleton set; it doesn't necessarily have to be Ord. Of course, you'll
find yourself unable to add other values to the set if it's not Ord, but
that's not `singleton`'s problem.

The empty redundancy makes sense as well, since an empty set can be a
set for any type, not necessarily ordered types. Ordering is only
required if the set may contain elements.

The `Eq` reduncancy is simply because `Ord` has already a constraint of
`Eq a => Ord a`.

This is true as of the current version:
https://hackage.haskell.org/package/base-4.8.2.0/docs/Data-Ord.html
and has been true since at least 4.0.0.0:
https://hackage.haskell.org/package/base-4.0.0.0/docs/Data-Ord.html