binary-search-tree: Allow empty lists/trees
This means bstValue has to be `BST a -> Maybe a` now, instead of its
current type of `BST a -> a`.

This impacts many tests that chained bstLeft or bstRight (both functions
of type `BST a -> Maybe (BST a)`) into bstValue. Whereas before you
would chain with fmap, now we must chain with monadic bind so that the
final type remains Maybe a (if we used fmap, it would be Maybe (Maybe a)).

The implementation has also gotten slightly more boilerplatey since we
don't get "free" implementations of bst{Value,Left,Right} anymore;
instead we have to explicitly consider the Tip and Node cases.

On the other hand, fromList is now simpler, since it can handle lists
that are either empty or not.

Closes #59.