binary-search-tree: Use Foldable in example toList
This doesn't affect the test interface at all, but makes sense in the
context of #59 - we can now fold over the tree using, say, the monoid of
lists and concatenation. This means we don't need to explicitly
implement `toList` but instead use the default implementation from
Data.Foldable and it will work!