binary-search-tree: Allow empty lists/trees
This means bstValue has to be `BST a -> Maybe a` now, instead of its
current type of `BST a -> a`.

This impacts many tests that chained bstLeft or bstRight (both functions
of type `BST a -> Maybe (BST a)`) into bstValue. Whereas before you
would chain with fmap, now we must chain with monadic bind so that the
final type remains Maybe a (if we used fmap, it would be Maybe (Maybe a)).

This is achieved by moving the old BST type to be a Node, and using a
newtype to define that a BST is a Maybe Node.

So, why not just use type instead of newtype? Simply because things like
the Foldable instance for Maybe don't make sense for BSTs.

The implementation has also gotten slightly more boilerplatey since we
don't get "free" implementations of bst{Value,Left,Right} anymore;
instead we have to explicitly fmap in the node versions of these functions.

Closes #59.