README: Explain significance of adding _test to path
Whether this is necessary depends on whom we are targeting with this
README. If we think everyone reading it already knows what it means to
add to the PATH, then we don't need this.

This commit, if we merge it, does not make that assumption, so we
explain why someone might want to add _test to path.

It's noted that the README doesn't provide instructions on how to add
something to the PATH but this would be a tricky endeavor as the way to
do so may differ, so we'd have to discuss it.