travis: test both Stack and non-Stack (cabal-install)
The work in #162 is advantageous because we can check our dependencies
and have hermetic builds (do not break on upstream changes).

One question about whether to merge it was the necessity to drop GHC
7.6.

This commit introduces to Travis the ability to test with either Stack
or without (the old way of doing things). We test 7.6, 7.8, 7.10, 8.0
with cabal-install as we did before, and also Stack as done in #162.

Closes #162 (is based on it)
Closes #124 (explanatory comments provided)
Closes #153 (based on #162 which also closed #153)