queen-attack: Explain positions and white/black
Taking the suggestions of https://github.com/exercism/xhaskell/pull/200/files#r71086758
into account, it seemed prudent to explain enough so that it is clear
from just looking at the stub file what is expected. This seems to be in
line with the goals of #181 which explains that we wish to save users
from inspecting the test suites to figure out what they need to
implement.