leap: Add variable to stub solution's signature.
Considering that `leap` is the first exercise of the track,
this small change may help a beginner, because he|she may
think that the part before `=` should not be changed.

More experienced students will certainly already know that
they can eta-reduce the function, so I think there is no
harm.