travis: Cache examples separately
PR #396 gives us the power to have multiple examples. But if the
examples have different dependencies, their caches clobber each other.
Thus, it seems necessary to have a separate cache per example.

There is some discussion on this and some proposed solutions in #397.

We may replace this shell script soon, but it will at least let us know
the strategy we want to use and validate the directory layout.