travis: Build examples in separate folders
Creating distinct folders to build each one of the examples,
we can be more confident that all tests are really independent.

Considering that we use distinct `.stack-work` directories,
the only possible contamination seems to be the global Stack's
cache. Now, with distinct paths for each example, it seems
improbable that this will be a problem.