travis: Upgrade to stack snapshot lts-7.9
We are using `nightly-2016-07-17` for too long, and the newer
snapshots have more packages and more updated versions.

Considering that now *nightly* and *lts* are both using
GHC-8.0.1, is makes sense to switch back to a *lts* snapshot.

This change will allow users to use more updated packages and
also possibly avoid some incompatibilities when using packages
that are not in *Stackage*.