difference-of-square: Comment out last test
The test `huge difference` has caused some difficulties
to users because the most common solutions use the `sum`
function, that seems to be implemented as a `foldr`.

A naive implementation would certainly fail the test or,
in the worst case scenario, allocate all the memory and
crash the operating system.