SETUP.md: Add for paranoia
In #462, we moved it to TRACK_HINTS.md

HOWEVER, we are the very first track to do so, so the code is untested.
It is possible something can go wrong.

If we want to be paranoid, we can add a SETUP.md so that if TRACK_HINTS
fails we don't leave our students in the dark.

This SETUP.md is slightly different from TRACK_HINTS.md: It has only one
space after the period on the last line, rather than two.

This way, we can tell which of the two files is being used.