travis: Build all stubs, and test all stubs except for untestable ones
In #421 we express a desire that all our stubs should compile.

Not all of them can successfully run their tests, so we will allow a
file that says not to run the tests. This file will be stored in the
.meta directory. Support for this was added in Trackler in
https://github.com/exercism/trackler/pull/4 - in particular, it will not
be served to students.

Regardless, all stubs should build, ensuring that we have not made a
syntax error!

If we desire, we can evaluate from time to time whether it is legitimate
for the exercises declaring DONT-TEST-STUB to remain that way.

We don't use `--pedantic` since there are many cases where, say, a
`Dummy` constructor is not used or a function argument is not used.

Closes #421