test-example, test-stub: clean up build directory after each
The build dir is not cached in Travis, so we can and should clean it up
locally. Otherwise, it could cause problems if files from a previous
build get included into a new build.

We put it in xhaskell to have it be in a consistent location: stack
checks whether the path has changed, so it needs to be in a stable path.