triangle: Mark divergence from x-common
If we followed https://github.com/exercism/x-common/pull/424 then we
would check boolean properties for triangles formed from the inputs. We
argue that this is unidiomatic in this language, so we are staying with
the old version of the tests. We mark the divergence as of this day, so
that the script will recognise this decision as having been made after
the x-common change.