Revert "pov: Add dummy symbolic link (#454)"
This reverts commit 860ee70daaa4da9893e9013b1517805104c16d7e.

In #442 we expressed the desire to use symlinks to avoid a duplicate
stack.yaml in each exercise.

However, RubyGems warns that symlinks are not supported on all
platforms, and there is a history of making incorrect link targets for
relative symlinks (see https://github.com/exercism/trackler/issues/13
for some details on that).

Worse, because of the incorrect symlink target, Trackler could not list
the files of the exercise, which prevented the README from being shown
(the page would instead show "Oops! That's our fault!" and no README).
See https://github.com/exercism/xhaskell/pull/507#issuecomment-287500563

Finally, in https://github.com/exercism/x-common/issues/601 we also
realize that symlinks may make it harder for contributors using
platforms without symlinks to contribute to this track.

Due to all these factors combined, we sadly give up on having symlinks.

Closes #442
Closes https://github.com/exercism/trackler/issues/13