travis: Remove error-type examples's cache
Exercise's examples that are always expected to
fail the building step are rebuilt by Stack
everytime. This triggers a cache update in
Travis-CI, even when nothing changed in the
repository.

While this problem doesn't affect pull requests
that already change the cache, it slows down
the all the other PRs by sometimes around 3 minutes.

Removing the cache folder for error-type examples,
we expect to have clean rebuilds in Travis.

Closes #544.