Add an explicit solution/ignore patterns to config
The default solution pattern matches more than just the reference solution.
It also matches some of the support files that live within the example directories, such
as yaml files.

The solution pattern should match the actual Haskell solution. This is
not used on the website per se, but I discovered the discrepancy when I
was scripting some pull requests to all the tracks, and the PR submitted
contained YAML rather than Haskell.

This makes the ignore pattern explicit, as we want to be clear that
we're not serving any files that have example in the path, even if
they are not the actual Haskell solution.