palindrome-product: Clarify requirements
It's not clear from the current examples what the requirements of this
exercise are.  In fact, a furtive glance is more likely to cause
confusion than give clarity.

In my opinion, this is because for the first example the solution lies
within the range of the list of factors, giving on a first impression
that we are looking for products that lie within this range.

I think that by listing out the entire list of products (which is small
in the case of the range of 1..9) we make it more clear that we are
considering all the results.

https://github.com/exercism/problem-specifications/pull/897