change 1.1.0.4: Test for greedy algorithm
This test invalidates an incorrect greedy algorithm which takes as many
of the larger coins as possible as long as remainder is greater than or
equal to the smallest coin. Currently such an algorithm will pass all of
the tests.

Unlike "possible change without unit coins available", at the point at
which you must stop using some coin C, the remainder still exceeds the
coin C + the smallest coin.

https://github.com/exercism/problem-specifications/pull/882