go-counting 1.0.0.3: Test two territories of same player
I requested this be added in
https://github.com/exercism/problem-specifications/pull/1195#pullrequestreview-99526421

There was an implicit test of this in the 5x5 board, but we'll start
small.

Note that this track deliberately does not change `territories` to be a
map mapping a player to all of their territories, instead keeping it an
array of all separate contiguous regions. This is because I (as the
track maintainer) believe that keeping it an array provides better
fidelity of the board and helps ensure that implementations are doing
the right thing.