go-counting 1.0.0.3
https://github.com/exercism/problem-specifications/pull/1195

Note that this track deliberately does not change `territories` to be a
map mapping a player to all of their territories, instead keeping it an
array of all separate contiguous regions. This is because I (as the
track maintainer) believe that keeping it an array provides better
fidelity of the board and helps ensure that implementations are doing
the right thing.