comment `library` section's `dependencies:` mapping
It has been observed that with Stack built against certain sets of
dependencies (exact details unknown), using operations such as `stack
test` would result in the following:

    package.yaml: Error while parsing $.library.dependencies - expected Array, Object, or String, encountered Null

This has been observed with Stack 1.6.5 in some circumstances, but not
in the Stack 1.6.5 at https://www.stackage.org/stack/linux-x86_64.

More details as to exactly which versions cause this are unknown because
an investigation was deemed unproductive.

This commit was automatically performed by running:

    sed -i 's/^  dependencies:/  # dependencies:/' */package.yaml

Note that hpack officially does not support null dependencies:

https://github.com/sol/hpack/blob/0.19.0/src/Hpack/Dependency.hs#L52
https://github.com/sol/hpack/blob/0.28.2/src/Hpack/Syntax/Dependency.hs#L60

It may be possible that some support structure around hpack was silently
dropping mapping keys with value null, but this does not seem like wise
behaviour to rely on.

Closes https://github.com/exercism/haskell/issues/672