clock 2.2.1.6: Remove clockHour and clockMin
Ever since the clock exercise was implemented for Haskell track in
https://github.com/exercism/haskell/pull/12, the tests have never
required clockHour nor clockMin; it was an internal function used by the
example solution.

when stub was added for clock (and other exercises) in July 2016 in
https://github.com/exercism/haskell/pull/200 the clockHour and clockMin
were added to the stub.

It doesn't seem to make sense to include untested functions in the stub,
plus there does not seem to be any particular value in testing these
separately (rather than the current approach of just testing toString),
so the solution is to remove them.