minesweeper: switch y and x coordinates
Otherwise, the example solution fails on any board that meets both
conditions:

* both dimensions larger than 1
* not symmetric on the diagonal formed by `(n, n)` forall n

For the test data used in the only such board is "large board"

To see why, see that `listArray (1, 1) (X, Y)` assigns indices to
elements in the order `(1, 1), (1, 2) ... (1, Y), (2, 1) ...`
However, we `concat` the `[String]` board, which means that successive
elements should increment the X coordinate, not the Y coordinate.
By switching X and Y in `listArray` (and all other uses), we get a
correct assignment.

Note that it's necessary for both dimensions to be larger than 1. If
either dimension is 1, only one of the dimensions gets increased in
`listArray`, therefore there is no danger of an incorrect assignment.

Solving the mystery of
https://github.com/exercism/haskell/pull/284