Remove date in package.yaml for non-versioned exercise
In #522 the following convention was established:

> [exercises based on a non-versioned canonical-data.json should use
> version 0.9.0.1. If there is a comment containing the date from the data
> file used as a reference, it should be added to the YAML file, to ease
> transition, like this:

The 'binary' exercise appears to be the only package left that honors
this convention of non-versioned canonical data. Since the exercise is
deprecated, and canonical data has versions in them, I propose that we
deprecate this convention in favor of 'version' lines that contain only
the MAJOR.MINOR.PATCH.SERIAL version number.