wordy 1.5.0.8: Test question `What is?` with no operands or operators
In keeping with [1.3.0][], we want to test that this fails in an
expected way rather than an unexpected way.

[1.3.0]: https://github.com/exercism/problem-specifications/pull/1383

Specifically for Haskell, we want to make sure that this results in
`None` rather than accessing an out-of-bounds index due to always
assuming that the inputs will have at least three space-separated
elements (of the form "What is X...?").

https://github.com/exercism/problem-specifications/pull/1401