pangram 1.4.1.10: Make stub Text-based
Pangram is the second string-heavy exercise with Bob being the first.

As part of the `Data.Text` initiative started in #760 and tracked in #780,
pangram will be the first exercise where the stub assumes the use of
`Data.Text` rather than `String`.

Because the Haskell ecosystem is still adapting, and the Prelude and
popular teaching material doesn't cover this extensively yet, we'll have
to provide some insight into why this is desirable.