pangram 1.4.1.9: Make test suite compatible with `Data.Text` (#780)
This mimics the change made for bob in #760.

The only thing that is necessary for the test suite to support a

    responseFor :: Text -> Text

is to assume that it has the type `IsString s => s -> s`.

We leave the stub as `String -> String`, but make the test suite
compatible with the change. This means that mentors can suggest a
solution based on `Data.Text` if they like.

A `Data.Text`-based solution is provided as an example.

Exercise version bumped from 1.4.1.8 to 1.4.1.9