pig-latin: Make test suite compatible with `Data.Text`
As in #760 (bob) and #788 (pangram), the only change necessary to support

    translate :: Text -> Text

is to enable OverloadedStrings in the test suite.

We leave the stub as `String -> String`, but make the test suite
compatible with the change, and we provide an optional README hint
adapted from #794 (bob). This may encourage students to pursue a
solution based on `Data.Text`, and it relieves mentors from having to
give this hint.

A `Data.Text`-based solution is provided as an example.

Exercise version bumped from 1.2.0.6 to 1.2.0.7.