isogram: Make test suite compatible with `Data.Text` (#780)
As part of #780, use `Data.Text` when appropriate.

The only change necessary to support

    isIsogram :: Text -> Bool

is to enable OverloadedStrings in the test suite.

We leave the stub as `String -> String`, but make the test suite
compatible with the change, and we provide an optional README hint
adapted from #805 (pig-latin). This may encourage students to pursue a
solution based on `Data.Text`, and it relieves mentors from having to
give this hint.

A `Data.Text`-based solution is provided as an example.

This example uses mono-traversable for monomorphic folding.

Exercise version bumped from 1.7.0.7 to 1.7.0.8.