acronym: Make test suite compatible with `Data.Text` (#780) (#811)
As part of #780, use `Data.Text` when appropriate.

And as in #760 (bob), #788 (pangram) and #805 (pig-latin), the only
change necessary to support

    abbreviate :: Text -> Text

is to enable OverloadedStrings in the test suite.

We leave the stub as `String -> String`, but make the test suite
compatible with the change, and we provide an optional README hint
adapted from #805 (pig-latin). This may encourage students to pursue a
solution based on `Data.Text`, and it relieves mentors from having to
give this hint.

A `Data.Text`-based solution is provided as an example.

Exercise version bumped from 1.6.0.9 to 1.6.0.10.