dnd-character: Remove name property
The name of the D&D character was provided as inspiration for populating
the `Character` type with additional fields that may be relevant in some
practical context.

But it doesn't serve a purpose, and the remaining abilities should also
serve as inspiration. Having seen a few solutions default the name to
`""`, the name should be removed in the context of this exercise.