etl: Make test suite compatible with `Data.Text` (#780)
The only change necessary to support

    transform :: Map Int Text -> Map Char Int

is to enable OverloadedStrings in the test suite.

We leave the stub as `Map a String -> Map Char a`, but make the test
suite compatible with the change, and we provide an optional README
hint adapted from #812 (isogram). This may encourage students to pursue
a solution based on `Data.Text`, and it relieves mentors from having to
give this hint.

A `Data.Text`-based solution is provided as an example.

Exercise version bumped from 1.0.0.4 to 1.0.0.5