Diamond: Implement property-based tests (#843)
Diamond's test suite is extended with a set of property-based tests.

These are in part inspired by the list of properties from README.md, and in
part by Mark Seemann's blog post on property-based checks for exactly this
exercise:

    https://blog.ploeh.dk/2015/01/10/diamond-kata-with-fscheck/

Only a subset of the properties mentioned are implemented.  This leaves room
for students to extend the test suite further, although exercism/exercism#5052
and #847 indicate that we may not wish to encourage this in the near future.

Because @barrymoo's property-based tests uncovered a bug in the example
solution, the example solution is replaced with one that passes current tests.