README: Remove 'Asking for help' section
The Exercism gitter chatroom is not Haskell-specific. There are people
qualified to answer Haskell questions who monitor the chat, but it's not
very active, and it's not being prioritized in exercism/discussions#43:

> The conclusion [in 2016] was to leave "as is" for now. (2017-06-05)

We choose to omit the following "chat on gitter" icon similar to the
"build passing" Travis CI icon, but keep it here for reference:

    [![Join the chat at https://gitter.im/exercism/support](https://badges.gitter.im/exercism/support.svg)](https://gitter.im/exercism/support?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge&utm_content=badge)

The "just ask" link is pointed to the issue tracker's "New issue".