diamond: Support Data.Text in test suite (#855)
Addresses #841.

Because the Diamond exercise uses property-based tests, it is not enough to 
assume that an instance of `IsString s` also has `Eq s`, since the properties
use several string-related functions. And because `String` is polymorphic and
`Text` is monomorphic, there is no common interface to these functions.

This os overcome with the 'string-conversions' package.