dominoes: added double-orientation test
I found a case where a student's code would mistakenly use the same
domino in both ways. With their code, the chain
`[(1, 2), (2, 3), (2,1)]` would be considered valid, as they would
first match (1,2) with (2,3), then swap the domino and match (3,2)
with (2,1).

Surprisingly, none of the tests currently catch this, hence this
patch.