Fix fetch-configlet. (#898)
* Fix fetch-configlet.

Travis builds are failing because of fetch-configlet failures, as
tracked in https://github.com/exercism/configlet/issues/173. This
patch is a copy of https://github.com/exercism/cpp/pull/344, which
fixed the cpp track.

* Applied suggested fix.

This makes the test case-insensitive rather than simply fixing the
case, by switching to perl. This is taken from
https://github.com/exercism/sml/commit/a721a5c525294bfdf41aef18ed83ffcbe83a08d6.