Rewrite "How to implement a Haskell concept exercise"
Before this commit, it appears that this document was a mixture between
a generic template and my initial writings about how we might go about
creating a framework for implementing concept exercises, i.e., not an
actual description of any agreed-upon approach to writing concept
exercises on the Haskell track.

Removes references to CIS 194.

Removes links to generic documents that don't seem useful.

Particularly: Preserve YouTube link of @iHiD talking. :)