Unbreak CI: Just include the entire history
I think 'with: ref: main' is causing the wrong checkout.

While the purpose of this is to check out the history far enough back to
compare against 'main' (formerly 'master'), it seems that we end up *on*
'main'?

Doing a 'git checkout HEAD' does not work because apparently HEAD refers
to main with 'with: ref: main'

In the absence of a better alternative, let's try with just including
the entire git history, meh.