crypto-square: Check whitespace
Fixes https://github.com/exercism/haskell/issues/895, the complaint that
the whitespace was being ignored.

The canonical test data itself is fine; the problem was that the test
suite was ignoring whitespace when comparing the actual result to the
expected result.

The decision to ignore whitespace was a originally deliberate decision,
with the reasoning in https://github.com/exercism/haskell/issues/259.

However, maintainers' current judgment is that as long as our output is
a String, it's better to keep it as the problem description says, for it
would be against the description if a student emitted an output with a
lot of leading spaces (among others).

If the track changes the output to a [String], then the maintainers
could freely choose whether to require the trailing spaces or forbid
them.