sgf-parsing: Add individual tests of escaping/whitespace behaviour
previous case "escaped property value" crams too much into one test
case.  Split it up into multiple ones.

The reimplemented case is mostly as it was when the exercise was
originally implemented in
https://github.com/exercism/exercism/commit/7a5075bf96e2cafb29f2f883ff98b708be5594a6

Note that the original case also got it wrong in that newlines should
remain newlines; this is corrected in the new case.

https://github.com/exercism/problem-specifications/pull/1889

https://github.com/exercism/haskell/pull/1025