Update installation instructions to use GHCup (#1101)
These days the recommended installation method is to use GHCup.

The previous instructions (based on Stack) were not wrong, but the GHCup
method is more general and should make it easier for students to do
Haskell stuff outside of Exercism. I don't think the instructions are
much more complicated.

I have test-driven (the GHCup-related bits of) these intructions only on Windows.